Remove unnecessary interface for DTO and fix usage
